+++ src/rerun/controller/RerunTabController.java
@@ -1,17 +1,5 @@
-/*
- * Copyright 2000-2015 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+/**
+ * Created by alutman on 21-Jan-15.
  */
 
 package rerun.controller;
@@ -29,6 +17,9 @@ import org.springframework.web.servlet.ModelAndView;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
 
 public class RerunTabController extends BaseFormXmlController {
 
@@ -51,18 +42,25 @@ public class RerunTabController extends BaseFormXmlController {
         Long buildId = Long.parseLong(request.getParameter("buildId"));
         String returnUrl = request.getHeader("Referer");
         try {
-            BuildPromotion buildPromotion = myServer.findBuildInstanceById(buildId).getBuildPromotion();
             SBuildType buildType = myServer.findBuildInstanceById(buildId).getBuildType();
 
-            BuildCustomizerFactory factory = myServer.findSingletonService(BuildCustomizerFactory.class);
+            Map<String, String> newParams = createParameterMap(
+                    request.getParameterValues("paramNames"), request.getParameterValues("paramValues")
+            );
 
             SUser user = SessionUser.getUser(request);
+            Iterator<Map.Entry<String, String>> iterator = buildType.getParameters().entrySet().iterator();
+            while(iterator.hasNext()) {
+                Map.Entry<String, String> next = iterator.next();
+                System.out.println(next.getKey() + " : " + next.getValue());
+            }
+
+            BuildCustomizerFactory factory = myServer.findSingletonService(BuildCustomizerFactory.class);
             BuildCustomizer customizer = factory.createBuildCustomizer(buildType, user);
-            customizer.setParameters(buildPromotion.getParameters());
+            customizer.setParameters(newParams);
 
             SQueuedBuild sQueuedBuild = customizer.createPromotion().addToQueue(user.getUsername());
             returnUrl = myServer.getRootUrl() + "/viewQueued.html?itemId=" + sQueuedBuild.getItemId();
-
             //TODO Send success or failure result
         } catch (Exception e) {
             e.printStackTrace();
@@ -72,5 +70,13 @@ public class RerunTabController extends BaseFormXmlController {
         } catch (IOException e) {
             e.printStackTrace();
         }
+    }
+
+    private Map<String, String> createParameterMap(String[] names, String[] values) {
+        Map<String, String> paramMap = new HashMap<String, String>();
+        for(int i = 0; i < names.length; i++) {
+            paramMap.put(names[i], values[i]);
+        }
+        return paramMap;
     }
 }